Refactor and add toggle for showing cluster
