+++ src/webapp/components/ClusterMap.js
@@ -1,8 +1,8 @@
-import React from "react";
+import React, { useState } from "react";
 import { Source } from "react-mapbox-gl";
 import { LngLat } from "mapbox-gl";
-import Alert from "@material-ui/lab/Alert";
-import { useTranslation } from "react-i18next";
+import Checkbox from "@material-ui/core/Checkbox";
+import FormControlLabel from "@material-ui/core/FormControlLabel";
 import { findBounds } from "../helpers/mapbox-coordinates";
 import {
   CROWN_HEIGHTS_BOUNDS,
@@ -11,6 +11,8 @@ import {
 import BasicMap from "./BasicMap";
 import { QuadrantsLayers } from "./QuadrantMap";
 import ClusterMapLayers from "./ClusterMapLayers";
+import { RequestNotFoundAlert, NoRequestsAlert } from "./MapAlerts";
+import { getRequestParam } from "../helpers/request-params";
 
 const makeBounds = (features) => {
   const lnglats = features.map((f) => {
@@ -26,59 +28,17 @@ const makeBounds = (features) => {
   return bounds;
 };
 
-const getRequestParam = () => {
-  const searchStr = window.location && window.location.search;
-  return searchStr
-    .slice(1)
-    .split("&")
-    .reduce((acc, token) => {
-      const matches = token.match(/request=(.*)/);
-      return matches ? matches[1] : acc;
-    }, "");
-};
-
-// Alert to show when a request from URL param does not exist
-const RequestNotFoundAlert = ({ requestCode }) => {
-  const { t: str } = useTranslation();
-  return (
-    <Alert severity="warning">
-      {`${str("webapp:deliveryNeeded.requestNotFound.message", {
-        defaultValue: `Request with code {{requestCode}} is not found. This means that the request is no longer in 'Delivery Needed' status.`,
-        requestCode,
-      })} `}
-      <a
-        href={str("webapp:deliveryNeeded.requestNotFound.redirectLink", {
-          defaultValue: "/delivery-needed",
-        })}
-      >
-        {str("webapp:deliveryNeeded.requestNotFound.redirectMessage", {
-          defaultValue: `See all requests instead.`,
-        })}
-      </a>
-    </Alert>
-  );
-};
-
-const NoRequestsAlert = () => {
-  const { t: str } = useTranslation();
-  return (
-    <Alert severity="warning">
-      {str("webapp:deliveryNeeded.noRequests.message", {
-        defaultValue:
-          "No requests found. Some requests may not have been posted in Slack yet or be marked for driving clusters.",
-      })}
-    </Alert>
-  );
-};
-
 const ClusterMap = ({ geoJsonData, containerStyle = {} }) => {
   const requestCode = getRequestParam();
+  const [showDrivingClusters, setShowDrivingClusters] = useState(true);
 
   let paramRequest;
   const { requests, drivingClusterRequests } = geoJsonData;
   const { features: reqFeatures } = requests;
   const { features: clusterFeatures } = drivingClusterRequests;
-  const allRequests = [...reqFeatures, ...clusterFeatures];
+  const allRequests = showDrivingClusters
+    ? [...reqFeatures, ...clusterFeatures]
+    : reqFeatures;
 
   if (requestCode) {
     // find first feature with code match to be passed
@@ -86,7 +46,7 @@ const ClusterMap = ({ geoJsonData, containerStyle = {} }) => {
     [paramRequest] = allRequests.filter(
       ({
         properties: {
-           meta: { Code },
+          meta: { Code },
         },
       }) => Code === requestCode
     );
@@ -103,6 +63,15 @@ const ClusterMap = ({ geoJsonData, containerStyle = {} }) => {
 
       {noRequestsFound && <NoRequestsAlert />}
 
+      <FormControlLabel
+        control={(
+          <Checkbox
+            checked={showDrivingClusters}
+            onClick={() => setShowDrivingClusters(!showDrivingClusters)}
+          />
+        )}
+        label="Show driving clusters"
+      />
       <BasicMap
         center={CROWN_HEIGHTS_CENTER_COORD}
         bounds={makeBounds(allRequests)}
@@ -134,11 +103,13 @@ const ClusterMap = ({ geoJsonData, containerStyle = {} }) => {
           paramRequest={paramRequest}
           color="orangered"
         />
-        <ClusterMapLayers
-          sourceId="drivingClusterRequestsSource"
-          paramRequest={paramRequest}
-          color="rebeccapurple"
-        />
+        {showDrivingClusters && (
+          <ClusterMapLayers
+            sourceId="drivingClusterRequestsSource"
+            paramRequest={paramRequest}
+            color="rebeccapurple"
+          />
+        )}
       </BasicMap>
     </>
   );
